requirements.txt
python-telegram-bot==13.7
google-cloud-dialogflow==2.9.1
vk-api==11.9.4
environs==9.3.4
